pomodoro-win11/requirements.txt
altgraph==0.17.4
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
playsound==1.2.2
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pywin32-ctypes==0.2.3
setuptools==80.9.0
